Added address finder to text component, some issues with registering window with redux-form
