+++ src/index.js
@@ -15,8 +15,8 @@ import Footer from './components/Footer';
 const store = createStore(combineReducers({
   reducers,
   form: formReducer
-// }), window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
-}), applyMiddleware(thunk));
+}), window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
+// }), applyMiddleware(thunk));
 
 class App extends React.Component {
   render(){
