+++ src/styles/Terms.css
@@ -1,27 +1,23 @@
 .terms {
   list-style: none;
   counter-reset: section; }
-
-.terms ol {
-  list-style: none;
-  padding-left: 0; }
-  .terms ol > li {
-    counter-increment: clause; }
-    .terms ol > li::before {
-      content: counter(clause,lower-alpha) ". "; }
-  .terms ol ol {
-    counter-reset: clause; }
-
-.terms li {
-  padding-left: 3rem;
-  margin: .5rem 0;
-  line-height: 1.25;
-  position: relative; }
-  .terms li::before {
-    position: absolute;
-    width: 2rem;
-    text-align: right;
-    left: 0; }
+  .terms ol {
+    list-style: none;
+    padding-left: 0; }
+    .terms ol > li {
+      padding-left: 3rem;
+      margin: .5rem 0;
+      line-height: 1.25;
+      position: relative;
+      counter-increment: clause; }
+      .terms ol > li::before {
+        content: counter(clause,lower-alpha) ". ";
+        position: absolute;
+        width: 2rem;
+        text-align: right;
+        left: 0; }
+    .terms ol ol {
+      counter-reset: clause; }
 
 .terms > li {
   counter-increment: section; }
