+++ src/styles/scss/App.scss
@@ -211,6 +211,7 @@ legend {
     opacity: 1;
     color: $font-color-primary;
   }
+  
   + div {
     background-color: $accordian-bg;
     margin-top: -30px;
